Move the pydevd hooks to their own module.
